+++ src/lib/components/disclosure/tab/TabAnchor.svelte
@@ -26,6 +26,10 @@
     } & IHTML5Properties &
         IGlobalProperties;
 
+    type $$Slots = {
+        default: {};
+    };
+
     export let element: $$Props["element"] = undefined;
 
     export let active: $$Props["active"] = false;