+++ src/lib/components/disclosure/tab/TabGroup.svelte
@@ -5,6 +5,10 @@
         logic_id: string;
     };
 
+    type $$Slots = {
+        default: {};
+    };
+
     export let logic_id: $$Props["logic_id"];
 
     const _logic_id = make_id_context(logic_id as string, CONTEXT_FORM_ID);