+++ src/lib/components/interactables/form/FormControl.svelte
@@ -17,6 +17,10 @@
         IGlobalProperties &
         IMarginProperties;
 
+    type $$Slots = {
+        default: {};
+    };
+
     export let element: $$Props["element"] = undefined;
 
     let _class: $$Props["class"] = "";