+++ src/lib/components/interactables/form/FormLabel.svelte
@@ -31,6 +31,10 @@
         IGlobalProperties &
         IPaddingProperties;
 
+    type $$Slots = {
+        default: {};
+    };
+
     export let element: $$Props["element"] = undefined;
 
     export let active: $$Props["active"] = false;