+++ src/lib/components/layouts/group/Group.svelte
@@ -18,6 +18,10 @@
         IMarginProperties &
         IPaddingProperties;
 
+    type $$Slots = {
+        default: {};
+    };
+
     export let element: $$Props["element"] = undefined;
 
     export let orientation: $$Props["orientation"] = undefined;