+++ src/lib/components/navigation/breadcrumb/BreadcrumbItem.svelte
@@ -15,6 +15,10 @@
         IGlobalProperties &
         IMarginProperties;
 
+    type $$Slots = {
+        default: {};
+    };
+
     export let element: $$Props["element"] = undefined;
 
     export let active: $$Props["active"] = false;