+++ src/lib/components/utilities/portal/Portal.svelte
@@ -12,6 +12,10 @@
         target?: HTMLElement | string;
     };
 
+    type $$Slots = {
+        default: {};
+    };
+
     const dispatch = createEventDispatcher();
 
     export let element: $$Props["element"] = undefined;